# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Samadhan AI — Grievance Redressal API
  version: 0.1.0
  description: |
    REST API for the Arunachal Pradesh State Grievance Redressal Platform.
    AI-powered complaint routing, multi-department RBAC, and citizen self-service.
  contact:
    name: Samadhan AI Team
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://your-app.vercel.app/api
    description: Production (Vercel)
...
paths:
...
  /complaints:
    get:
      tags: [Complaints]
      summary: List complaints (admin, paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [pending, triage, in_progress, resolved, closed, escalated] } }
        - { name: priority, in: query, schema: { type: string, enum: [low, medium, high, critical] } }
        - { name: search, in: query, schema: { type: string }, description: Full-text search }
        - { name: department, in: query, schema: { type: string } }
        - { name: assignedTo, in: query, schema: { type: string } }
        - { name: slaBreached, in: query, schema: { type: string, enum: ['true', 'false'] } }
        - { name: dateFrom, in: query, schema: { type: string, format: date } }
        - { name: dateTo, in: query, schema: { type: string, format: date } }
        - { name: sort, in: query, schema: { type: string, default: '-createdAt' } }
        - { name: cursor, in: query, schema: { type: string }, description: MongoDB _id for cursor-based pagination }
      responses:
        '200':
          description: Paginated complaint list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/Complaint' } }
                  meta:
                    type: object
                    properties:
                      page: { type: integer }
                      limit: { type: integer }
                      total: { type: integer }
                      totalPages: { type: integer }
                      nextCursor: { type: string, nullable: true }

    post:
      tags: [Complaints]
      summary: Submit a new complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, submitterName, submitterPhone]
              properties:
                title: { type: string, minLength: 5, maxLength: 200 }
                description: { type: string, minLength: 20, maxLength: 5000 }
                submitterName: { type: string }
                submitterPhone: { type: string }
                submitterEmail: { type: string, format: email }
                location: { type: string }
                state: { type: string }
                district: { type: string }
                coordinates: { type: object, properties: { lat: { type: number }, lng: { type: number } } }
                callConsent: { type: boolean, default: false }
                attachments: { type: array, items: { type: object } }
      responses:
        '201': { description: Complaint created with tracking ID (GRV-XXXX) }
        '429': { description: Rate limited }
...
  /complaints/{id}/notes:
    get:
      tags: [Complaints]
      summary: List internal notes
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Array of notes }
    post:
      tags: [Complaints]
      summary: Add internal note
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '201': { description: Note created }
...
components:
  schemas:
    Complaint:
      type: object
      properties:
        _id: { type: string }
        complaintId: { type: string, description: 'Tracking ID (GRV-AP-XXXX-NNNN)' }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        priority: { type: string, enum: [low, medium, high, critical] }
        status: { type: string, enum: [pending, triage, in_progress, resolved, closed, escalated] }
        department: { type: string }
        location: { type: string }
        state: { type: string }
        district: { type: string }
        aiSummary: { type: string, nullable: true }
        aiConfidence: { type: number, nullable: true }
        slaDeadline: { type: string, format: date-time, nullable: true }
        slaBreached: { type: boolean }
        analysisStatus: { type: string, enum: [queued, processing, completed, deferred] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT access token (admin)
    citizenCookieAuth:
      type: apiKey
      in: cookie
      name: citizen_access_token
      description: JWT access token (citizen)
    cronSecret:
      type: http
      scheme: bearer
      description: Vercel CRON_SECRET for cron endpoints